ADDING SSH COMPONENT [log] printf support %.*s; [stdport] anneal std socket api; impl ntoh/hton_xxx; add unistd.h; [task] add sshd_task(); [crypto] add crypto files: chacha, crypto_api, hash, off_t, poly1305, random, sha2, smult_curve_xxx [esp,sock] changed sock_recv behavior: do not clear buff; [ssh] add session main: acpt_loop.c; add basics: context, packet, packet_def; add kexinit part; [vo] fix dump unrecog %z; [user] switch demo to sshd_task

+++ User/types/vo.c
@@ -153,7 +153,7 @@ void vbuff_dump(const vstr_t *str) {
 void buff_dump(const void *data, uint32_t len) {
   const uint8_t *p = data;
   for (size_t i = 0; i < len; i += 16) {
-    printf("%.4zu:", i);
+    printf("%.4u:", (int)i);
     for (size_t j = i; j < i + 16; j++) {
       if (j < len)
         printf("%02x ", p[j]);
